+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.ts
@@ -1,14 +1,22 @@
 import angular from "angular";
+import noUiSliderModule from 'angularjs-nouislider';
 import "../../core/core.module";
-import {horizontalMetricChooserComponent, metricChooserComponent} from "./metricChooser.component";
+import {
+    areaMetricChooserComponent,
+    colorMetricChooserComponent,
+    heightMetricChooserComponent
+} from "./metricChooser.component";
 
-angular.module("app.codeCharta.ui.metricChooser",["app.codeCharta.core"]);
+angular.module("app.codeCharta.ui.metricChooser",["app.codeCharta.core", noUiSliderModule]);
 
 angular.module("app.codeCharta.ui.metricChooser").component(
-    metricChooserComponent.selector,
-    metricChooserComponent
+    areaMetricChooserComponent.selector,
+    areaMetricChooserComponent
 ).component(
-    horizontalMetricChooserComponent.selector,
-    horizontalMetricChooserComponent
+    heightMetricChooserComponent.selector,
+    heightMetricChooserComponent
+).component(
+    colorMetricChooserComponent.selector,
+    colorMetricChooserComponent
 );