+++ visualization/app/codeCharta/ui/optionsPanel/optionsPanel.component.ts
@@ -1,0 +1,21 @@
+import {SettingsService} from "../../core/settings/settings.service";
+import "./optionsPanel.component.scss";
+
+export class OptionsPanelController {
+
+    /* @ngInject */
+    constructor(
+        private settingsService: SettingsService
+    ) {
+    }
+
+}
+
+export const optionsPanelComponent = {
+    selector: "optionsPanelComponent",
+    template: require("./optionsPanel.component.html"),
+    controller: OptionsPanelController
+};
+
+
+